Rubocop changes. Increased Limits for Complexity
Reasoning:
OAuth defines a lot of errors to returned to the client upon the first sign of problems.
These are not trivially handled without branches in the program. CyclomaticComplexity
and PerceivedComplexity do however count each occurence of tokens like `||`,`if`, etc.
and throw errors for each function having too many of those.

Why wasn't this a problem so far?
Most error handling was done within blocks of Sinatras `get` or `post` functions,
which do not count as functions.

Why is this a problem now?
As routines get more complicated due to new implemented standards, the code would become
unreadable without implementing some behaviors in their own functions.
Also, we might want to reuse code for some endpoints adressable via both `get` and `post`
in the future.